Refactor to named exports only for handler
It appears to have been a major refactor to the internal implmentation of serverless-offline between 5.x to 6.x.

The latest version does not appear to handle default exports correctly.

Pre 6.x the `export default` in hello/index.js resulted in the following object being required:

  {
    hello: [AsyncFunction: o],
    default: [AsyncFunction]
  }

Where as the latest series it miss-parses once required by the InProcessRunner.js handler runner as:

  {
    hello: [AsyncFunction: o],
    default: { hello: [AsyncFunction: o], default: [AsyncFunction] }
  }

This resulted in serverless-offline throwing an expception* as it expects a function but instead got an object
(default: {...} above)

To resolve this for the time being I have switched the hanlder definition to be explicit and removed the export
default from the handler.

This required a slight change as to when the usage of runWarm wrapper.

I have elected to use this approach for now to resolove the issue pending review as I'm not entirely sure this
is not related to babel?

* Exception reported by serverless-offline
handler 'default' in '/home/rd/Development/serverless-starter/.webpack/service/functions/hello/index' is not
a function