Update unit test for hello function
Inline usage of the runWarm wrapper requires a slight change to the unit test, passing an event object. This is used by the wrapper to determine if the event source is aws.

This was not an issue before as it was used on the default export which had no test coverage.