Pin serverless related package vesions
Lots of npm i errors related to node-gyp (or other) dependancy on nan

Only way to resolve this was to rebuild the package lock file as it was in a state where dependencies where not updating correctly.

Due to the chevron versions it installed latest serverless rather than 1.67.3 thus pint exact versions.