Add "package" script
It can be useful to package functions locally (rather than just
deploying straight to the provider). The Serverless CLI tool provides
this ability, this just adds an npm script to make it a little easier to
use in line with the version of the Serverless package installed as a
dependency.